Added sealed traits to make the API more clear and put database behind a crate feature to allow using of the timezone utilities in the library instead of pulling the entire IANA db.
